+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Era.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -6,12 +8,17 @@ module Test.Cardano.Ledger.Dijkstra.Era (
   DijkstraEraTest,
 ) where
 
+import Cardano.Ledger.Block (Block (..))
 import Cardano.Ledger.Dijkstra (DijkstraEra)
+import Cardano.Ledger.Dijkstra.Era (DijkstraEraBlockHeader (..))
 import Cardano.Ledger.Dijkstra.Scripts (DijkstraEraScript)
 import Cardano.Ledger.Dijkstra.State
 import Cardano.Ledger.Dijkstra.TxBody (DijkstraEraTxBody)
 import Cardano.Ledger.Plutus (Language (..))
 import Data.Coerce
+import Data.Maybe (fromJust)
+import Lens.Micro (lens)
+import Test.Cardano.Ledger.BlockHeader (TestBlockHeader (..))
 import Test.Cardano.Ledger.Conway.Era
 import Test.Cardano.Ledger.Dijkstra.Arbitrary ()
 import Test.Cardano.Ledger.Dijkstra.Binary.Annotator ()
@@ -45,3 +52,9 @@ instance BabbageEraTest DijkstraEra
 instance ConwayEraTest DijkstraEra
 
 instance DijkstraEraTest DijkstraEra
+
+instance DijkstraEraBlockHeader TestBlockHeader DijkstraEra where
+  blockHeaderPrevNonceL =
+    lens
+      (fromJust . tbhPrevNonce . blockHeader)
+      $ \b@Block {blockHeader} pNonce -> b {blockHeader = blockHeader {tbhPrevNonce = Just pNonce}}