+++ eras/shelley/test-suite/bench/Cardano/Ledger/Shelley/Bench/Gen.hs
@@ -10,6 +10,7 @@ module Cardano.Ledger.Shelley.Bench.Gen (
   genChainState,
 ) where
 
+import Cardano.Ledger.Block (EraBlockHeader)
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Shelley.API (
   ApplyBlock,
@@ -29,6 +30,7 @@ import Control.State.Transition.Extended
 import Data.Either (fromRight)
 import qualified Data.Map.Strict as Map
 import Data.Proxy
+import Test.Cardano.Ledger.BlockHeader (TestBlockHeader)
 import Test.Cardano.Ledger.Shelley.BenchmarkFunctions (ledgerEnv)
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (MockCrypto)
 import Test.Cardano.Ledger.Shelley.Constants (
@@ -50,8 +52,6 @@ import Test.Cardano.Ledger.Shelley.Rules.Chain (ChainState (..))
 import Test.Cardano.Ledger.Shelley.Serialisation.Generators ()
 import qualified Test.Control.State.Transition.Trace.Generator.QuickCheck as QC
 import Test.QuickCheck (generate)
-
--- ===============================================================
 
 -- | Generate a genesis chain state given a UTxO size
 genChainState ::
@@ -83,7 +83,8 @@ genBlock ::
   , GetLedgerView era
   , EraRule "LEDGERS" era ~ ShelleyLEDGERS era
   , QC.HasTrace (ShelleyLEDGERS era) (GenEnv MockCrypto era)
-  , ApplyBlock era
+  , ApplyBlock TestBlockHeader era
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   GenEnv MockCrypto era ->
   ChainState era ->