+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rules/Chain.hs
@@ -6,7 +6,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
@@ -28,7 +27,6 @@ module Test.Cardano.Ledger.Shelley.Rules.Chain (
   chainStateNesL,
 ) where
 
-import Cardano.Ledger.BHeaderView (BHeaderView)
 import Cardano.Ledger.BaseTypes (
   BlocksMade (..),
   Nonce (..),
@@ -36,7 +34,7 @@ import Cardano.Ledger.BaseTypes (
   StrictMaybe (..),
  )
 import Cardano.Ledger.Binary (EncCBORGroup)
-import Cardano.Ledger.Block (Block (..))
+import Cardano.Ledger.Block (BbodySignal (..), Block (..), EraBlockHeader)
 import Cardano.Ledger.Chain (
   ChainPredicateFailure (..),
   chainChecks,
@@ -82,7 +80,6 @@ import Cardano.Protocol.TPraos.BHeader (
   bheaderBlockNo,
   bheaderSlotNo,
   lastAppliedHash,
-  makeHeaderView,
   prevHashToNonce,
  )
 import Cardano.Protocol.TPraos.Rules.Prtcl (
@@ -254,7 +251,7 @@ instance
   , Embed (EraRule "BBODY" era) (CHAIN era)
   , Environment (EraRule "BBODY" era) ~ BbodyEnv era
   , State (EraRule "BBODY" era) ~ ShelleyBbodyState era
-  , Signal (EraRule "BBODY" era) ~ Block BHeaderView era
+  , Signal (EraRule "BBODY" era) ~ BbodySignal era
   , Embed (EraRule "TICKN" era) (CHAIN era)
   , Environment (EraRule "TICKN" era) ~ TicknEnv
   , State (EraRule "TICKN" era) ~ TicknState
@@ -268,6 +265,7 @@ instance
   , AtMostEra "Alonzo" era
   , State (EraRule "LEDGERS" era) ~ LedgerState era
   , EraCertState era
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   STS (CHAIN era)
   where
@@ -289,7 +287,7 @@ chainTransition ::
   ( Embed (EraRule "BBODY" era) (CHAIN era)
   , Environment (EraRule "BBODY" era) ~ BbodyEnv era
   , State (EraRule "BBODY" era) ~ ShelleyBbodyState era
-  , Signal (EraRule "BBODY" era) ~ Block BHeaderView era
+  , Signal (EraRule "BBODY" era) ~ BbodySignal era
   , Embed (EraRule "TICKN" era) (CHAIN era)
   , Environment (EraRule "TICKN" era) ~ TicknEnv
   , State (EraRule "TICKN" era) ~ TicknState
@@ -303,6 +301,7 @@ chainTransition ::
   , State (EraRule "LEDGERS" era) ~ LedgerState era
   , EraGov era
   , EraCertState era
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   TransitionRule (CHAIN era)
 chainTransition =
@@ -317,7 +316,7 @@ chainTransition =
                    etaC
                    etaH
                    lab
-               , Block bh txs
+               , blk@(Block bh _)
                )
            ) -> do
         case prtlSeqChecks lab bh of
@@ -326,7 +325,6 @@ chainTransition =
 
         let pp = nes ^. nesEpochStateL . curPParamsEpochStateL
             chainChecksData = pparamsToChainChecksPParams pp
-            bhView = makeHeaderView bh Nothing
 
         -- We allow one protocol version higher than the current era's maximum, because
         -- that is the way we can get out of the current era into the next one. We test
@@ -334,7 +332,7 @@ chainTransition =
         -- to the next era. This, of course, works properly only with HFC (Hard Fork
         -- Combinator), which is implemented in consensus, but for the purpose of the
         -- CHAIN test this is OK.
-        case chainChecks (succ (eraProtVerHigh @era)) chainChecksData bhView of
+        case chainChecks (succ (eraProtVerHigh @era)) chainChecksData blk of
           Right () -> pure ()
           Left e -> failBecause (RealChainPredicateFailure e)
 
@@ -369,7 +367,7 @@ chainTransition =
 
         BbodyState ls' bcur' <-
           trans @(EraRule "BBODY" era) $
-            TRC (BbodyEnv pp' account, BbodyState ls bcur, Block bhView txs)
+            TRC (BbodyEnv pp' account, BbodyState ls bcur, BbodySignal blk)
 
         let nes'' = updateNES nes' bcur' ls'
             bhb = bhbody bh